test artifact 2

+++ build_logic/src/main/kotlin/telegram/tasks/DetailInfoTask.kt
@@ -63,9 +63,7 @@ abstract class DetailInfoTask @Inject constructor(
                             chatId = chatId.get()
                         )
                     } finally {
-                        if (tempReportFile.exists()) {
-                            tempReportFile.delete()
-                        }
+                        if (tempReportFile.exists()) tempReportFile.delete()
                     }
                 }
             }
